added some design to different pages

+++ server/app/routes/product_routes.py
@@ -12,11 +12,12 @@ def get_products():
     products = Product.query.all()
     return jsonify([p.to_dict() for p in products])
 
-@product_bp.route('/', methods=['POST'])
+@product_bp.route('/add_products', methods=['POST'])
 @jwt_required()
 def create_product():
     user_id = get_jwt_identity()
-    user = User.query.get(user_id)
+    user = User.query.get(int(user_id))
+
 
     if user.role != "admin":
         return jsonify({"error": "Admin only"}), 403
@@ -29,6 +30,11 @@ def create_product():
 
     if not all([name, price, description, brand, image]):
         return jsonify({"error": "Missing fields"}), 400
+    
+    try:
+        price = float(price)
+    except ValueError:
+        return jsonify({"error": "Price must be a number"}), 422
     
     result = cloudinary.uploader.upload(image)
     image_url = result["secure_url"]
@@ -47,7 +53,14 @@ def create_product():
     return jsonify(product.to_dict()), 201
 
 @product_bp.route("/<int:id>", methods=["DELETE"])
+@jwt_required()
 def delete_product(id):
+    user_id = get_jwt_identity()
+    user = User.query.get(int(user_id))
+
+    if user.role != 'admin':
+        return jsonify({"error": "Admin only"}), 403
+
     product = Product.query.get_or_404(id)
     db.session.delete(product)
     db.session.commit()
